// frontend/metropolitan/src/app/shared/model/nota-debito-calculate.ts
export interface NotaDebitoCalculate {
  codCliente: number;
  codOperador: number;
  codCounter: number;
  codTipoCambio: number;
  typeCounter: number;
  totalArgentina: number;
  totalAgencia: number;
  totalCounter: number;
  totalMetropolitana: number;
  total: number;
  montoNeto: number;  
}